backend/requirements.txt
asgiref
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
bcrypt
requests
regex
Pillow